refactor(loader): relocate computeWorldMatrices helper from GLTFMetalKit to GLTFCore
Relocates the generic glTF parent-child world-matrix accumulation algorithm from the Metal rendering layer (GLTFMetalKit) to the foundational parsing layer (GLTFCore).

- refactor(loader): move `computeWorldMatrices(document:into:)` and `localMatrix(for:)` to new `GLTFSceneGraph` enum inside GLTFCore/Animation/SceneGraph.swift.
- refactor(loader): expose `simd_float4x4` translation and scale initializers as public extensions of `simd_float4x4` in GLTFCore so they are reusable across downstream modules.
- refactor(loader): update GLTFAssetLoader's matrix pre-pass to call `GLTFSceneGraph.computeWorldMatrices` directly, avoiding code duplication and exposing the primitive for future glTF flat-traversal consumers in VRMMetalKit.
